cvorg: added support to feed the PLL with an external clock.
The implemented algorithm to find the proper values has one handicap:
it finds good values for all the parameters except the 'r'. The PLL
doesn't lock.

Forcing 'r' equals 500, the algorithm calculates the other values and
the PLL is locked. The beneath problem seems to be a restriction in the
value of this parameter but in the datasheet was not found anything.
This is a work-around for the specific application of our client.

The client will feed the CVORG's PLL with an external clock of 10 MHz (PS
reference).

Signed-off-by: Samuel Iglesias Gonsalvez <siglesia@cern.ch>