cvorb driver: change cvorb_t handler structure
We have been request to add debugging level per device
To this end and to face any future extensions, the opaque
handler returned by teh open has been modified in a backkward comptible
way (size stays unchanged). The goal is to have a pointer to private
data that we can easily extend to face future extension.