cvora: fix vmeio_riob_s size bug
Declaring vmeio_riob_s with variable size depending on __64BIT results
in the kernel expecting different ioctl numbers than the ones offered by
the library, because __64BIT is undefined in user space. The way it is
used here does not make any good.

So a single declaration of the structure for all the architectures is
kept. We choose to keep the compat field, given that, by historical
accident (and by bug) it is the one chosen in all architectures, and,
incidentally, in the one where this is proven to work (32-bit) and with
clients already running.

Libraries need not be modified, so no version pushing, full
compatibility at the ioctl() level is kept, etc.