sis33: fix installation of SIS3320-250
The installation scripts for sis3320 modules assume a 1:1
correspondence between module type (name) and driver. This is not
the case, as the sis3320 driver handles both SIS3320 and SIS3320-250
modules on an equal basis (actually, their LUN space is shared).

The problem arises when running

vmedesc $* --format=flat SIS3320

to obtain insmod parameters from transfer.ref, because vmedesc does
not accept a pattern, but an exact name. So, while SIS3320 and
SIS3320-250 should be accepted and equally treated, this is not the
case.

Although patching vmedesc to accept a regexp for the module type might
seem the answer, I prefer not to do so, because

1. We want to get rid of Perl in installation scripts for maintenance
reasons.
2. We do the same kind of parsing everywhere, nowadays, with shell
and/or python.
3. vmedesc does not parse irq vectors/levels, because this was not part
of the transfer.ref ioconfig stanzas at the time of its writing.

I believe these are sufficient reasons to replace soon vmedesc by a
solution that does the same (and more) based on the current
transfer.ref format and relying on something non-Perl.

Hence, I furnish this horrible patch to handle the situation of the
SIS3320-250 modules (currently just 3 instances), without impacting the
current SIS3320 installations (neither the vmods that rely equally on
vmedesc), in the hope that a vmedesc replacement will be deployed during
the 15th June technical stop.