sis33: lib: add internal trigger support
This changeset adds the *_internal_trigger_* part of the API to
the sis33 library. It never worked properly (the hardware makes
a lousy job at trigger detection), and will not work without the
corresponding changes in the driver. This is provided only to
allow compilation of old sources relying on these definitions.