vmebridge/driver: make local copy of kernel's function pci_find_parent_resource
Linux kernel's function pci_find_parent_resource has changed in commit:
f44116ae PCI: Remove pci_find_parent_resource() use for allocation
This commit changed functionality of this function in a way that it cannot
be used anymore. Due to this there is a local copy of old version of the
function.

Function vme_bridge_init uses pci_find_parent_resource to find PCI bridge's
resource with flag IORESOURCE_MEM. By not specifying range of resource, old
version of the function returns a range of allocated region. However, new
function expects the range of resource to contain subrange of PCI bridge.
The problem is that PCI bridge resource mapping is not known at this point.
Even more this is the information that we're looking for.

As a alternative checking of PCI bridge resource range could be removed.
In this case later the allocation function shall report errors.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>