vmod:[fix] lun_to_dev() must check all device instances
Problem:
if you declare the first vmod module with LUN=2 (or any number grater than 0)
then the driver will not load the mezzanines.

The real device instance is stored in device_table according to its LUN.
This means that we have to loop over all the table to find the correct
instance. Before the condition was the device instances counter.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>