keep old Perl installation script
This was probably Emilio's first attempt at a Perl installation script
for sis33 modules. It is kept here for historical reasons, but it is
dubious that it has been run since 2010.

Dropping this commit is safe if no news about sis33 installs are heard
after Feb 2016, or the non-Perl-based install scripts are deployed.