cvorbtest: fix off-by-one bug in FSM display
The range of FSM states in the firmware is inclusive up to
CVORB_CH_MAX_STATES; prior to this fix, the last state was reported as
"unknown".

In addition, truly unknown states are now reported and displayed for
informational purposes. These "unknown" states are actually reached in
some cases.

This bugfix must be rebased upstream when the whole build system is
merged into master