driver: match devices by ID from user input or CSR
We keep compatibility with the platform_data matching mechanism. For
all new driver there will be the possibility to match using the
board ID (vendor:device).

The user can provide this information throught the sysfs attribute
`register_device`. If the user does not specify an ID the vmebus will
try to get it from the CSR space of VME64x modules.

Signed-off-by: Federico Vaga <federico.vaga@cern.ch>